FolderDB and ScripturePickerWindow now only look at files marked by RB as visible.  Fixes issue where SPW would bomb trying to read a hidden file that was first in the directory listing.
Corrected typo in SongML that caused a capo chord on a single chord line to print using the chord font, not the capo font.  Also fixed a spacing issue in the same routine.

Add the 128x128 program icon to the tracked files.

Turn off the "zoom button" attribute on ScripturePicker and modified the MacProcID to fix issues on Windows and Mac.  Windows shouldn't allow minimize, and Mac now allows resize.

git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@103 7d629990-da18-0410-9c13-aedeb0f471ca