Song selection window only opens once if activated immediately after entering presenter mode and no song is inserted. PresentWindow.KeyDown returns the result of KeyDownX and if the popup returned Nil the return was False, in other words, keystroke not handled.  Just got sent right back at KeyDownX somehow.
SongPickerWindow list of songs changes the highlighted song name with Page Up/Page Down/Home/End keys in keeping with commonly accepted UI practices. [Bug 1559333]  Added RowHeight, PageDown and PageUp extensions to the ListBox in OpenSongUtils.

ActionSongPrint and ActionSongPrintAll now pass the song file's path to LeadSheet.AddSong.  This clears up an issue where the highest file in the folder hierarchy would be used regardless of the one selected if there are duplicate filenames in different folders. (Prepend CellTag to the filename.)


git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@117 7d629990-da18-0410-9c13-aedeb0f471ca