Fix in Bible to trap Nil return from Xql method.
Fix in Scripture Picker Window to remove two Close statements that were not the last statements being executed.  This caused SPW to die with a NilObjectException caused by a method that was still active at the close trying to modify a control (which was Nil).

Convert ActionAskSetSave to use new SaveDialog.  Added a variable to track pop_sets_sets current item so the user can return to the last active set if Cancel is  selected after changing the selected set.  Rippled changes down to methods that call ActionAskSet Save.

Reworked ActionAskSongSave to handle all save & revert actions within that method rather than having to replicate this multiple times in the calling methods.  When ActionAskSongSave (or Set) returns, the in-memory state reflects what the user requested, with True/False indicating Continue/Cancel.

Two new undocumented flags to indicate if a Don't Save/Cancel will revert or stay with the current document's state.

git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@132 7d629990-da18-0410-9c13-aedeb0f471ca