Combined the changes for save/don't save/cancel for songs and sets from the trunk and Joshua's branch.  The result is a hybrid of both.
Updated SmartML.GetValueB to set the value to the default if both Create and a default are specified and the XmlNode doesn't exist.  While the previous logic would work if SetValueB was never called, the resultant XML was inaccurate if the default value was True.

Moved save of preferences to App.Close from MainWindow.Close where it more properly belongs.

git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@138 7d629990-da18-0410-9c13-aedeb0f471ca