English translation: edit text for CCLI license and CCLI song number to make the distinction clearer.
Add SetML.StyleChangeActive method.  Determines if the base style has been overridden and returns True if so.  This is used to fix a bug in BL15 where the wrong style would be used if PresentWindow.Present was called from the song editor and the active song had a custom style.

Fix mouse getting stuck on the hourglass after presentation. (PresentWindow.Close)

Fix PresentWindow on Mac so the resize box does not appear. As a temporary side effect, the window cannot be sized or moved when using dual-screen preview mode. (PresentWindow ProcID property)

Mouse cursor properly changes when over PresentWindow even if the window does not have focus. (PresentWindow.MouseEnter and MouseExit)

Corrected error in saving custom slides where the starting characters of the final slide would be dropped if the slides contained multi-byte characters. (OpenSongUtils.SplitRE) (Bug 1668194)

git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@154 7d629990-da18-0410-9c13-aedeb0f471ca