-= This revision is V1.0.0 BL 17=-
FolderDB delays recursing down the folder hierarchy until after the current folder is complete.  The status display is much clearer this way.

(App) Fix issue where OpenSong would quit if the documents folder saved in the preferences did not exist.  Now prompt the user like there is no folder in preferences.

(preferences plist) Add AbsolutePath to the info saved with a FolderItem in a preferences file.  OpenSong ignores this value when reloading the FolderItem, but it is useful for debugging.

(ScripturePickerWindow) When new scripture version is selected, don't reload the scripture if the new version is the same as the currently loaded version.

(ScripturePickerWindow) Release Bible object before loading new version.  Improves, but does not fully correct, a memory leak when changing bibles. [Bug 1703561]

Correct an issue introduced in r166 on IndexProgress where it wouldn't close properly after the scripture index is written. Also set IndexProgress window properties to prevent resizing.

(SongML) Made printing the presentation order on the leadsheet optional.  The default is to not print it and produce the same output as RC1-BL14. <leadsheet presentation="true"/> must be put in PrintSettings file by manual edit to enable.  Post-V1 will address this a different way; for now it provides a workaround for those who don't need or want the presentation order printed.


git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@167 7d629990-da18-0410-9c13-aedeb0f471ca