Incorporated Martyn's Set Export functionality based on the work in r215-231 in his private branch.
Fixed issue with the new FolderDB returning a FolderItem instead of Nil if the file does not exist.

Disabled CCLI Import.

Doing a GetValue on a plist-based pref died if an intermediate leaf in the tree didn't already exist. The missing leaves are now added.

Cleaned up some issues with FileUtils, especially missing error messages in the translation files.