Major reworking of the source trunk:
1. Corrected some lingering line ending issues in text files.

2. All text files properly labeled with SVN properties to use the native end-of-line style.  This _shouldn't_ create a problem with the program since line endings are converted internally if needed to a common (Unix LF) format.

3. Rebuilt the .rbvcp source files using the V1.2.1 OpenSong.xml source.  After significant research into the rbvcp format and the changes that have been made since RB2007r2, the build problem was discovered to be a result of a change to the way bound controls are identified in VCP format files.

The .rbvcp files were created by RB2007r5, but will successfully compile on versions at least as far back as r2.  Syntax for OpenSong source is still frozen at 2007r1.  By using 2007r5, it appears the consistency of the files will be better for many of the edit/merge operations we do.

There are no new features or bugfixes introduced in this revision.



git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@276 7d629990-da18-0410-9c13-aedeb0f471ca