Incorporates Joshua's changes through r273.  This includes the start of the activity log and several bug fixes.  See "changelog.txt" for full details.
Change IsCMMClick (deprecated keyword) to IsContextualClick throughout.  Only the name changed to protect the innocent. ('Dragnet' fans, unite!)

Fixed RB VCP format source (.rbvcp, .rbfrm and .rbbas files, etc.).  It will compile in versions of the IDE < 2007r4 again.  For reference, the bindings (grep "#tag Bindings") are done by name in >=2007r4 and by control number in <2007r4.  These need to be refactored out of OpenSong anyway since the bindings feature is deprecated in RB as of 2007r5.

Bumped version to V1.3.0 and StageCode to "Beta".


git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@277 7d629990-da18-0410-9c13-aedeb0f471ca