Update version strings and build number for V1.3b2
Enable tag replacement in changelog and put $Date$ and $Rev$ tags at bottom of file.  Rationale: I forgot to put a tag branch on Beta 1 and this will make it easier to figure these things out in the future.  Since changelog gets updated for nearly every commit to the trunk, it should always have the current revision for the build.


git-svn-id: https://opensong.svn.sourceforge.net/svnroot/opensong/source/trunk@310 7d629990-da18-0410-9c13-aedeb0f471ca