- Fix for multiscreen setup in Linux. Displays are now properly detected with support for different resolutions per screen (Bug #1904753) The code contains a workarround for the fact that OpenSong is not properly reporting number of screens and individual resolutions on Linux. This workarround requires multiple target specific implementations which is not nice. This workarround is still kept as fallback, but in general OpenSong will now use a custom class OS_Screen which replaces the default Screen class. The OS_Screen class supports proper screen size detection (using Gdk API). See the notes in the OS_Screen class.
- Fix for fullscreen issue in Linux where title bar remains visible (Feature request #2916937)
The described workarround is to make the window resizable, but this requires manual interaction and is not preserved over a presentation restart. This fix implements window layout modification hiding the title bar and window borders for the presentation window (using Gtk API).

Both issues are in the tracker and discussed in this forum thread: http://sourceforge.net/projects/opensong/forums/forum/373378/topic/4040032